:sparkles: (Home) refactor scroll analytics logic
